requirements.txt
Flask==2.2.5
flask_wtf==1.2.2
matplotlib==3.5.1
numpy==1.21.5
opencv_contrib_python==4.9.0.80
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==9.0.1
Pillow==11.0.0
psycopg2_binary==2.9.9
PyPDF2==3.0.1
pytesseract==0.3.13
pyzbar==0.1.9
WTForms==3.2.1
